==> 1 Data exploration commentary.Rmd <==
---
title: "Data exploration and commentary"
author: "Paul Hively"
date: "March 29, 2016"
output: html_document
---

Source can be viewed [on GitHub](https://github.com/phively/gsb-sols-proj/blob/master/1%20Data%20exploration%20commentary.Rmd).

# Problem statement

My goal is to find a way to accurately forecast fundraising revenue.

The existing model ("JMod") has two components. First, the probability a solicitation $S_{i}$ comes in before the end of the fiscal year is a function of the current stage progress $c$ and time until the end of the year $t$, $P(S_{i}=1)=f(c_{i},t)$. The expected FRP in the current fiscal year is this probability times the expected amount, $E(FRP)=P(S_{i}=1)E(S_{i})$.

Focus on the probability model. I assume that there should also be some effect due to the average close rate, season (e.g. people give more in December), planned ask amount, actual ask amount, and potentially interactions. Something like this:

$$P(S_{ij}=1)=f(c_{i},t,\mu,t_{i}^{*},a_{j}^{*},a_{j})$$

It makes sense to start with a few straightforward classifiers, but I'd also like to look into nesting models.

# Visualizations

The dataset includes all closed (successful = booked, or unsuccessful = not booked) solicitations with an add date on or after 7/1/2011. Data is obtained from the first tab of the "Booth solicitation history" report (saved as a .csv).

```{r, message=F, cache=F}
#### Run Rscript0 to load useful packages and functions ----
source("Rscript0 - libraries.R")
source("f.Wrangle.R")
#### Run Rscript1 to load data ----
source("Rscript1 - data load and transform.R")
```

* [Rscript0 - libraries.R](https://github.com/phively/gsb-sols-proj/blob/master/Rscript0%20-%20libraries.R)
* [f.Wrangle.R](https://github.com/phively/gsb-sols-proj/blob/master/f.Wrangle.R)
* [Rscript1 - data load and transform.R](https://github.com/phively/gsb-sols-proj/blob/master/Rscript1%20-%20data%20load%20and%20transform.R)

Begin with some visualizations of time in each stage:

```{r, echo=F, message=F, warning=F, cache=F}
load("modeling.Rdata")

#### Days between stages ----
ggdat <- Joe.dat2 %>%  mutate(
  plan2clear = as.numeric(difftime(Clear.Dt, Planning.Dt, units="days")),
  clear2ask = as.numeric(difftime(Ask.Dt, Clear.Dt, units="days")),
  ask2oral = as.numeric(difftime(Oral.Dt, Ask.Dt, units="days")),
  ask2actual = as.numeric(difftime(Actual.Dt, Ask.Dt, units="days")),
  oral2actual = as.numeric(difftime(Actual.Dt, Oral.Dt, units="days")),
  plan2actual = as.numeric(difftime(Actual.Dt, Planning.Dt, units="days")),
  giftsize = factor(ifelse(Expected.Amt>=5000000, 5000,
                    ifelse(Expected.Amt>=1000000, 1000,
                    ifelse(Expected.Amt>=100000,100,0))),
  labels=c("Under $100k","$100k+","$1M+","$5M+"))) %>%
# Only use closed solicitations; Stage 10=Booked, 12=Refused, 14=Cancelled, etc.
  filter(as.numeric(substr(Final.Sol.Stage,1,2)) >= 10)

## Create plots
bins <- 40
ggplot(ggdat, aes(x=plan2clear, fill=giftsize)) + geom_histogram(aes(y=..density..), alpha=.5, bins=bins) + geom_density(alpha=.5) + facet_grid(giftsize~.) + labs(title="Days from planning to clearance") + xlim(c(0,1000))
ggplot(ggdat, aes(x=clear2ask, fill=giftsize)) + geom_histogram(aes(y=..density..), alpha=.5, bins=bins) + geom_density(alpha=.5) + facet_grid(giftsize~.) + labs(title="Days from clearance to ask")
ggplot(ggdat, aes(x=ask2actual, fill=giftsize)) + geom_histogram(aes(y=..density..), alpha=.5, bins=bins) + geom_density(alpha=.5) + facet_grid(giftsize~.) + labs(title="Days from ask to actual")
ggplot(ggdat, aes(x=plan2actual, fill=giftsize)) + geom_histogram(aes(y=..density..), alpha=.5, bins=bins) + geom_density(alpha=.5) + facet_grid(giftsize~.) + labs(title="Days from planning to actual")
```

There's extremely obvious bimodality in most cases. Note in particular that while PG (defined as $5M+) is skewed positive overall, there seem to be two distinct groups: takes a while and takes a *long* while. The purple "Days from ask to actual" is about as perfect a mixture of two $N(\mu_{k},\sigma^{2}_{k})$ as I've ever seen. Might be interesting to see if a certain population just takes longer in general.

```{r, echo=F, warning=F, cache=F}
## Ask and oral dates
bins <- 40
ggplot(ggdat[ggdat$giftsize!="$5M+",], aes(x=ask2oral, fill=giftsize)) + geom_histogram(aes(y=..density..), alpha=.5, bins=bins) + geom_density(alpha=.5) + facet_grid(giftsize~.) + labs(title="Days from ask to oral")
ggplot(ggdat[ggdat$giftsize!="$5M+",], aes(x=oral2actual, fill=giftsize)) + geom_histogram(aes(y=..density..), alpha=.5, bins=bins) + geom_density(alpha=.5) + facet_grid(giftsize~.) + labs(title="Days from oral to actual")
```

There isn't much data in the oral pledge stage so things are a bit bumpier here.

Here's the cleaned up data file:
```{r, echo=T, warning=F, cache=F}
#### Model data cleanup ----
source("Rscript1a - modeling data appends.R")
colnames(mdat)
```

* [Rscript1a - modeling data appends.R](https://github.com/phively/gsb-sols-proj/blob/master/Rscript1a%20-%20modeling%20data%20appends.R)

**Booked** indicates whether the solicitation came in at $>$$0, and **FY.plan, FY.clear, FY.ask, FY.oral** indicate whether the solicitation closed in the same fiscal year as that stage was reached, while the corresponding **FY.[Stage].Book** indicates that a solicitation was booked in the same fiscal year that it reached that stage. Combinations of these indicators will be the dependent variables. **plan2actual, clear2actual, ask2actual, oral2actual** are measured in days. **Planned.Dt** is the planned *ask* date, while **Planning.Dt** is the date the solicitation entered the planning stage. The other covariates should be self-explanatory.

# Variable exploration

## Dependent variables

Do I want to model time to close from each current stage $c_{i}$ as opposed to $P(S_{ij}=1)$?

```{r, echo=F, warning=F, cache=F}
ggdat <- gather(mdat, "Stages","Days", 19:22, factor_key=T) #gather() uses tidyr
ggplot(data=ggdat, aes(x=Stages, y=Days, color=Booked, fill=Booked)) + geom_boxplot(alpha=.4, position=position_dodge(width=1))
kable(summary(mdat[,19:22]))
```

Negative values indicate a data issue or update. Decidedly right-skewed. Try a square root transformation:

```{r, echo=F, warning=F, cache=F}
ggplot(data=ggdat, aes(x=Stages, y=sqrt(Days), color=Booked, fill=Booked)) + geom_boxplot(alpha=.4, position=position_dodge(width=1))
```

Not bad. How's log look?

```{r, echo=F, warning=F, cache=F}
ggplot(data=ggdat, aes(x=Stages, y=log(Days), color=Booked, fill=Booked)) + geom_boxplot(alpha=.4, position=position_dodge(width=1))
```

Might be a good idea to try e.g. Box-Cox transformations to see how they compare. 

Data counts - do I have enough to do what I want? Is it remotely reasonable to split by month or quarter?

```{r, echo=F, warning=F, cache=F}
## Group data by month
tmp <- mdat %>% mutate(Plan=as.numeric(month(Planning.Dt)),
                       Clear=as.numeric(month(Clear.Dt)),
                       Ask=as.numeric(month(Ask.Dt)),
                       Oral=as.numeric(month(Oral.Dt)), Count=1) %>%
  select(Count, Plan, Clear, Ask, Oral, Booked)

# Convert columns into rows
tmp2 <- gather(tmp, "Stage", "Month", 2:5, factor_key=T) %>% mutate(Month=factor(Month, labels=strtrim(month.name, 3)))
tmp2$Month <- factor(tmp2$Month, levels(tmp2$Month)[c(7:12,1:6)]) #start in Jul instead of Jan
ggdat <- tmp2 %>% group_by(Stage, Month) %>% summarise(Count=sum(Count))
# Plot
ggplot(na.omit(ggdat), aes(x=Month, y=Count, color=Stage, fill=Stage, group=Stage)) + geom_point() + geom_line() + labs(y="Count of solicitations entering stage") + theme(text=element_text(size=12))
# Table
kable(xtabs(Count ~ Stage + Month, data=ggdat))
```

Peak ask occurs around Nov-Dec and Mar each year. Not nearly as much oral pledge data; usually oral to booked is fast enough that the date isn't captured in the system. (JMod rolls ask and oral together.)

Finally, let's take a look at the booked versus declined solicitations by stage by fiscal year closed.

```{r, echo=F, warning=F, cache=F}
## Plot count of solicitations booked by month and stage
ggdat <- na.omit(tmp2)
ggplot(ggdat, aes(x=Month, color=Booked, fill=Booked)) + geom_bar(alpha=.4) + labs(y="Count of solicitations entering stage") + theme(text=element_text(size=12)) + facet_grid(Stage~.)
## Proportions
ggdat <- na.omit(tmp2) %>% group_by(Stage, Month, Booked) %>% summarise(Count=sum(Count))
tmp3 <- na.omit(tmp2) %>% group_by(Stage, Month) %>% summarise(Count=sum(Count))
# Convert each count in ggdat to a percent
for(i in 1:dim(ggdat)[1]){
  ggdat$Prop[i] <- ggdat$Count[i]/(tmp3$Count[tmp3$Stage==ggdat$Stage[i] & tmp3$Month==ggdat$Month[i]])
  # Binomial CIs, se^2 = p(1-p)/n
  ggdat$ci[i] <- sqrt((ggdat$Prop[i]*(1-ggdat$Prop[i]))/tmp3$Count[tmp3$Stage==ggdat$Stage[i] & tmp3$Month==ggdat$Month[i]])
}
## Plot
ggplot(ggdat, aes(x=Month, y=Prop, color=Booked, fill=Booked, group=Booked)) + geom_point() + geom_line(alpha=.4) + geom_ribbon(aes(ymin=Prop-ci, ymax=Prop+ci), alpha=.15) + labs(y="Proportion of solicitations entering stage") + theme(text=element_text(size=12)) + facet_grid(Stage~.) + scale_y_continuous(labels=percent)
```

Confidence bands are based on the variance for a binomial random variable, $\sqrt{\frac{p(1-p)}{n}}$ (assuming independence).

This is fascinating; a solicitation entering plan or clear in January or plan in April-July is likely to be refused rather than booked. Is there a specific year driving that?

```{r, echo=F, warning=F, cache=F}
## Redo previous plots for each fiscal year
tmp <- mdat %>% select(Planning.Dt, Clear.Dt, Ask.Dt, Oral.Dt, Booked) %>%
  mutate(Count=1, PlanY=YToFY(Planning.Dt), Plan=as.numeric(month(Planning.Dt)),
         ClearY=YToFY(Clear.Dt), Clear=as.numeric(month(Clear.Dt)),
         AskY=YToFY(Ask.Dt), Ask=as.numeric(month(Ask.Dt)),
         OralY=YToFY(Oral.Dt), Oral=as.numeric(month(Oral.Dt))) %>%
  select(-c(1:4)) #Drop actual dates which are now unneeded
# Spread month and year
tmp2 <- gather(tmp, "Stage", "Month", seq(4,10,by=2), factor_key=T) %>% gather("StageY", "Year", 3:6) %>%
  mutate(Month=as.factor(Month), Year=as.factor(Year)) %>%
  # Need to dedupe to get rid of mismatched rows, e.g. where Stage=Ask and StageY=PlanY
  filter((Stage=="Plan" & StageY=="PlanY") | (Stage=="Clear" & StageY=="ClearY") | (Stage=="Ask" & StageY=="AskY") | (Stage=="Oral" & StageY=="OralY")) %>% mutate(StageY=NULL)
levels(tmp2$Month) <- strtrim(month.name, 3) #use month names
tmp2$Month <- factor(tmp2$Month, levels(tmp2$Month)[c(7:12,1:6)]) #start in Jul instead of Jan
## Redo previous # plot for each fiscal year
ggdat <- na.omit(tmp2) %>% group_by(Stage, Month, Year) %>% summarise(Count=sum(Count))
ggplot(ggdat, aes(x=Month, y=Count, color=Stage, fill=Stage, group=Stage)) + geom_point() + geom_line() + labs(y="Count of solicitations entering stage by year") + facet_grid(Year~.) + theme(text=element_text(size=12))
```

Note that this is fiscal, not calendar, year. FY2011 didn't have much stage progress, as it is the first year of data.

```{r, echo=F, warning=F, cache=F}
## Data for plotting %s
ggdat <- na.omit(tmp2 %>% group_by(Stage, Month, Year, Booked) %>% summarise(Count=sum(Count)))
tmp3 <- na.omit(tmp2 %>% group_by(Stage, Month, Year) %>% summarise(Count=sum(Count)))
# Convert each count in ggdat to a percent
for(i in 1:dim(ggdat)[1]){
  ggdat$Prop[i] <- ggdat$Count[i]/(tmp3$Count[tmp3$Stage==ggdat$Stage[i] & tmp3$Month==ggdat$Month[i] & tmp3$Year==ggdat$Year[i]])
    # Binomial CIs, se^2 = p(1-p)/n
  ggdat$ci[i] <- sqrt((ggdat$Prop[i]*(1-ggdat$Prop[i]))/tmp3$Count[tmp3$Stage==ggdat$Stage[i] & tmp3$Month==ggdat$Month[i] & tmp3$Year==ggdat$Year[i]])
}
## Plot
ggplot(ggdat[!(ggdat$Year %in% c("2011","2016")),], aes(x=Month, y=Prop, color=Booked, fill=Booked, group=Booked)) + geom_point() + geom_line(alpha=.4) + geom_ribbon(aes(ymin=Prop-ci, ymax=Prop+ci), alpha=.15) + labs(y="Proportion of solicitations entering stage") + facet_grid(Stage~Year) + scale_y_continuous(labels=percent) + theme(axis.text.x=element_text(angle=90, vjust=.4))
```

Looks like 2013 was a particularly good year, but it's normal to have lots of duds in those winter months

## Discrete covariates

I'm thinking time to close is not the natural outcome measure; a model predicting time to close can't be swapped out directly with the corresponding terms in JMod. Recall from above:

> For the probability model, I assume that there should also be some effect due to the average close rate, season (e.g. people give more in December), planned ask amount, actual ask amount, and potentially interactions. Something like this:
>
> $$P(S_{ij}=1)=f(c_{i},t,\mu,t_{i}^{*},a_{j}^{*},a_{j})$$

Take a look at each of these additional variables against Booked.

### Ask Amount

```{r, echo=F, warning=F, cache=F}
bins <- 5
g1 <- ggplot(mdat, aes(x=Ask.Amt, y=..count.., color=Booked, fill=Booked)) + geom_histogram(alpha=.2, bins=bins, position="identity") + geom_density(alpha=.5) + labs(x="Ask amount", y="Count", title=expression(paste("Close counts by ",Log[10]," ask amount"))) + scale_x_continuous(trans="log10", breaks=10^seq(1:7), labels=scales::dollar) #+ facet_grid(Booked~.)
# Look at relative frequencies
g2 <- ggplot(mdat, aes(x=Ask.Amt, y=..count.., color=Booked, fill=Booked)) + geom_hline(yintercept=.5, alpha=.5) + geom_density(alpha=.5, position="fill") + labs(x="Ask amount", y="Probability", title=expression(paste("Close rates by ",Log[10]," ask amount"))) + scale_x_continuous(trans="log10", breaks=10^seq(1:7), labels=scales::dollar)
# Final layout
grid.arrange(g1, g2)
```

Histogram and smoothed density estimate. It's above 50% close rate at every level, but the poorest close rate is between the $100k to $1M levels. There's obviously nonlinearity; a cubic polynomial might be a good approximation?

### Ask Amount/Planned Amount

I think differences in planned ask versus actual ask amounts might indicate trouble. How does the ratio of actual ask over planned ask look?

```{r, echo=F, warning=F, cache=F}
g1 <- ggplot(mdat, aes(x=(Ask.Amt/Planned.Amt), y=..count.., color=Booked, fill=Booked)) + geom_density(alpha=.5) + labs(x="Ask amount", y="Count", title="Close counts by Ask/Planned") + scale_x_continuous(trans="log10", labels=scales::percent)
# Look at relative frequencies
g2 <- ggplot(mdat, aes(x=(Ask.Amt/Planned.Amt), y=..count.., color=Booked, fill=Booked)) + geom_hline(yintercept=.5, alpha=.5) + geom_density(alpha=.5, position="fill") + labs(x="Ask amount", y="Proportion", title="Close rates by Ask/Planned") + scale_x_continuous(trans="log10", labels=scales::percent)
# Final layout
grid.arrange(g1, g2)
```

Too choppy to really see much of anything; most solicitations are asked at the planned amount and there are so few in the tails that there isn't much interesting to say. What if we just look at a few binned thresholds?

```{r, echo=F, warning=F, cache=F}
# Custom binning
bins <- c(0,.25,.5,.75,.99,1.01,1.5,2,2.5,3)
## If I keep copy-pasting this it should really go in its own .R script
g1 <- ggplot(mdat, aes(x=Ask.Amt/Planned.Amt, color=Booked, fill=Booked)) + geom_histogram(alpha=.5, breaks=bins, position="identity") + labs(x="Ask amount", y="Count", title="Close counts by Ask/Planned") + scale_x_continuous(labels=scales::percent, breaks=c(seq(0,1,by=.25), seq(1.5,max(bins),by=.5)))
# Look at relative frequencies
g2 <- ggplot(mdat, aes(x=Ask.Amt/Planned.Amt, color=Booked, fill=Booked)) + geom_hline(yintercept=.5, alpha=.5) + geom_histogram(alpha=.5, position="fill", breaks=bins) + labs(x="Ask amount", y="Proportion", title="Close rates by Ask/Planned") + scale_x_continuous(labels=scales::percent, breaks=c(seq(0,1,by=.25), seq(1.5,max(bins),by=.5)))
# Final layout
grid.arrange(g1, g2)
```

If someone asks slightly more than the planned ask amount it's a good sign that the talks were going well; otherwise there doesn't seem to be much to see here. The 25%-50% bin might also be statistically significant; don't know about the others given the small sample sizes.

## Stage progression within a single year

Given that a solicitation reaches a certain stage, how likely is it to close that year, whether or not it's booked?

```{r, echo=F, warning=F, cache=F}
ggdat <- mdat %>% mutate(Count=1) %>% group_by(FY.plan, FY.clear, FY.ask, FY.oral) %>% summarise(Count=sum(Count)) %>% gather("Stage","Closed",1:4,factor_key=T) %>% group_by(Stage, Closed) %>% summarise(Count=sum(Count)) %>% na.omit()
levels(ggdat$Stage) <- c("Plan","Clear","Ask","Oral")
# Plot of counts
ggplot(ggdat, aes(x=Stage, y=Count, fill=Closed)) + geom_bar(stat="identity", position="dodge", alpha=.5)
# Plot of probabilities
ggdat <- ggdat %>% mutate(n=sum(Count), Proportion=Count/n, ci=sqrt(Proportion*(1-Proportion)/n))
ggplot(ggdat, aes(x=Stage, y=Proportion, color=Closed, fill=Closed)) + geom_point(stat="identity", alpha=.5, size=2) + geom_linerange(aes(ymin=Proportion-ci, ymax=Proportion+ci), alpha=.25, size=3) + labs(y="Proportion") + scale_y_continuous(limits=c(0,1), breaks=seq(0,1,by=.2))
kable(ggdat[,c("Stage","Closed","Count","Proportion")], digits=2)
```

Ask and Clear are closer than I would have thought. What about if we look at whether they were both closed _and_ booked that year?

```{r, echo=F, warning=F, cache=F}
ggdat <- mdat %>% mutate(Count=1) %>% group_by(FY.Plan.Book, FY.Clear.Book, FY.Ask.Book, FY.Oral.Book) %>% summarise(Count=sum(Count)) %>% gather("Stage","Booked.In.FY",1:4,factor_key=T) %>% group_by(Stage, Booked.In.FY) %>% summarise(Count=sum(Count)) %>% na.omit()
levels(ggdat$Stage) <- c("Plan","Clear","Ask","Oral")
# Plot of counts
ggplot(ggdat, aes(x=Stage, y=Count, fill=Booked.In.FY)) + geom_bar(stat="identity", position="dodge", alpha=.5)
# Plot of probabilities
ggdat <- ggdat %>% mutate(n=sum(Count), Proportion=Count/n, ci=sqrt(Proportion*(1-Proportion)/n))
ggplot(ggdat, aes(x=Stage, y=Proportion, color=Booked.In.FY, fill=Booked.In.FY)) + geom_point(stat="identity", alpha=.5, size=2) + geom_linerange(aes(ymin=Proportion-ci, ymax=Proportion+ci), alpha=.25, size=3) + labs(y="Proportion") + scale_y_continuous(limits=c(0,1), breaks=seq(0,1,by=.2))
kable(ggdat[,c("Stage","Booked.In.FY","Count","Proportion")], digits=2)
```

Curious; there really doesn't seem to be a discernible difference between clear and ask. The last thing to check is these hybrid "reached stage and closed" indicators by month.

```{r, echo=F, warning=F, cache=F}
# Data manupulation
ggdat <- na.omit(mdat %>% mutate(Plan=as.numeric(month(Planning.Dt)), Clear=as.numeric(month(Clear.Dt)), Ask=as.numeric(month(Ask.Dt)), Oral=as.numeric(month(Oral.Dt)), Count=1) %>% select(Count, Plan, Clear, Ask, Oral, FY.Plan.Book, FY.Clear.Book, FY.Ask.Book, FY.Oral.Book) %>% gather("Stage", "Month", 2:5, factor_key=T) %>% gather("Outcome", "Booked.In.FY", 2:5) %>% 
  # Need to get rid of mismatched rows, e.g. where Stage=Ask and Outcome=FY.Clear.Book
  filter((Stage=="Plan" & Outcome=="FY.Plan.Book") | (Stage=="Clear" & Outcome=="FY.Clear.Book") | (Stage=="Ask" & Outcome=="FY.Ask.Book") | (Stage=="Oral" & Outcome=="FY.Oral.Book")) %>% mutate(Month=factor(Month, labels=strtrim(month.name, 3))) %>% group_by(Stage, Month, Booked.In.FY) %>% summarise(Count=sum(Count)))
# Reorder months to be in fiscal order
ggdat$Month <- factor(ggdat$Month, levels(ggdat$Month)[c(7:12,1:6)])
# Add proportions and CIs
ggdat <- ggdat %>% mutate(n=sum(Count), Prop=Count/n, ci=sqrt(Prop*(1-Prop)/n))
# Plot counts by month
ggplot(ggdat, aes(x=Month, y=Count, color=Booked.In.FY, group=Booked.In.FY)) + geom_point() + geom_line(alpha=.5) + facet_grid(Stage~.) + theme(text=element_text(size=12))
# Plot prop by month
ggplot(ggdat, aes(x=Month, y=Prop, color=Booked.In.FY, fill=Booked.In.FY, group=Booked.In.FY)) + geom_point() + geom_line(alpha=.5) + geom_ribbon(aes(ymin=Prop-ci, ymax=Prop+ci), alpha=.15) + labs(y="Percent") + theme(text=element_text(size=12)) + facet_grid(Stage~.) + scale_y_continuous(labels=percent)
```

There are a couple of crossover points; January for Plan and Clear; Jan for ask (though it crosses back in March interestingly). Can't count on oral pledges made in May or June coming in before the end of the year. Let's focus on asks in the MG range, $x\in[\$25\text{k},\$5\text{M})$:

```{r, echo=F, warning=F, cache=F}
# Data manupulation
ggdat <- na.omit(mdat %>% mutate(GiftSize=ifelse(Ask.Amt>=5000000,"PG",ifelse(Ask.Amt>=25000,"MG",ifelse(Ask.Amt<25000,"AF","Unasked")))) %>% mutate(Plan=as.numeric(month(Planning.Dt)), Clear=as.numeric(month(Clear.Dt)), Ask=as.numeric(month(Ask.Dt)), Oral=as.numeric(month(Oral.Dt)), Count=1) %>% select(Count, Plan, Clear, Ask, Oral, FY.Plan.Book, FY.Clear.Book, FY.Ask.Book, FY.Oral.Book, GiftSize) %>% gather("Stage", "Month", 2:5, factor_key=T) %>% gather("Outcome", "Booked.In.FY", 2:5) %>% 
  # Need to get rid of mismatched rows, e.g. where Stage=Ask and Outcome=FY.Clear.Book
  filter((Stage=="Plan" & Outcome=="FY.Plan.Book") | (Stage=="Clear" & Outcome=="FY.Clear.Book") | (Stage=="Ask" & Outcome=="FY.Ask.Book") | (Stage=="Oral" & Outcome=="FY.Oral.Book")) %>% mutate(Month=factor(Month, labels=strtrim(month.name, 3))) %>% group_by(GiftSize, Stage, Month, Booked.In.FY) %>% summarise(Count=sum(Count)))
# Reorder months to be in fiscal order
ggdat$Month <- factor(ggdat$Month, levels(ggdat$Month)[c(7:12,1:6)])
# Add proportions and CIs
ggdat <- ggdat %>% mutate(n=sum(Count), Prop=Count/n, ci=sqrt(Prop*(1-Prop)/n))
# Plot counts by month
ggplot(ggdat, aes(x=Month, y=Count, color=Booked.In.FY, group=Booked.In.FY)) + geom_point() + geom_line(alpha=.5) + facet_grid(Stage~GiftSize) + theme(text=element_text(size=12), axis.text.x=element_text(angle=90, vjust=.4))
# Plot prop by month
ggplot(ggdat, aes(x=Month, y=Prop, color=Booked.In.FY, fill=Booked.In.FY, group=Booked.In.FY)) + geom_point() + geom_line(alpha=.5) + geom_ribbon(aes(ymin=Prop-ci, ymax=Prop+ci), color=NA, alpha=.15) + labs(y="Percent") + theme(text=element_text(size=12), axis.text.x=element_text(angle=90, vjust=.4)) + facet_grid(Stage~GiftSize) + scale_y_continuous(labels=percent)
```

Ha, that's more like it. Plan has a negative trend after Feb; Clear is surprisingly close to flat; Ask has a negative trend after March; Oral can't be counted on in May and June.

# Packages used
```{r}
session_info()
```
